requirements.txt
streamlit
openai
Pillow 